fix: useVRMExpressionの関数を細分化
