+++ frontend/src/features/VRM/VRMExpression/types.ts
@@ -1,0 +1,12 @@
+export type ExpressionPreset =
+	| "neutral"
+	| "happy"
+	| "sad"
+	| "angry"
+	| "surprised"
+	| "relaxed";
+
+export type VRMExpressionState = {
+	preset: ExpressionPreset;
+	weight: number;
+};
